initial commit

tsc passing


fixes


fixes


working


fixups


updates


fixes


fixes


fmt
