+++ packages/codemirror-lsp-client/.gitignore
@@ -1,0 +1,6 @@
+node_modules
+build
+dist
+tsconfig.tsbuildinfo
+*.d.ts
+*.js
