+++ packages/codemirror-lsp-client/src/client/codec/bytes.ts
