+++ packages/codemirror-lsp-client/src/client/codec/headers.ts
