+++ packages/codemirror-lsp-client/src/client/server-capability-registration.ts
